pyproject.toml
[project]
name = "docusearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.60.1",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0,<0.4.0",
    "langchain-docling>=1.1.0",
    "langchain-huggingface>=0.1.2",
    "langchain-milvus>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "sentence-transformers>=5.1.2",
    "tiktoken>=0.12.0",
    "python-dotenv>=1.0.0",
    "pymilvus[milvus_lite]>=2.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
